rag_system/RAG/requirements.txt
pinecone
sentence-transformers
groq
python-dotenv
pytesseract
pdf2image
pdfminer.six
pillow
python-docx
python-pptx
numpy
torch
transformers
